Refactor project structure and add Jest configuration for testing #1
