+++ jest.config.ts
@@ -1,0 +1,9 @@
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "jsdom",
+  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
+  transform: {
+    "^.+\\.(ts|tsx)$": "ts-jest",
+  },
+  setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
+};
