+++ src/features/Keyboard.tsx
@@ -46,7 +46,7 @@ export default function Keyboard({
     <div
       style={{
         display: "grid",
-        gridTemplateColumns: "repeat(auto-fit, minmax(75px, 1fr))",
+        gridTemplateColumns: "repeat(auto-fill, minmax(75px, 1fr))",
         gap: "0.5rem",
       }}
     >
